requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
sqlmodel
psycopg2-binary
aiofiles
python-multipart
jose
bcrypt
passlib
pypdf
python-docx
requests
yt-dlp
youtube-transcript-api
moviepy
